backend/pyproject.toml

[tool.poetry]
name = "syncar-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
celery = "^5.3.6"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
playwright = "^1.40.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
